chore(devtools-ui): move reload-on-changes component
